+++ packages/components/src/components/table/collector.ts
@@ -6,6 +6,9 @@ import {
   useRefWeakMap,
   useRefWeakSet,
   UseExpandMethods,
+  UseSelectMethods,
+  UseSelectState,
+  useSelectMethods,
 } from '@lun-web/core';
 import { TableSetupProps, TableColumnSetupProps } from './type';
 import { getCollectorOptions } from 'common';
@@ -27,13 +30,25 @@ export type TableProvideExtra = {
   showResize(target: HTMLElement, column: InternalColumn): void;
   virtual: ReturnType<typeof useVirtualList>;
   rowExpand: UseExpandMethods;
+  select: ReturnType<typeof useSelectMethods>;
   state: TableState;
 };
+
+export type TableExternalContext = {
+  rowExpand: UseExpandMethods;
+  rowSelect: UseSelectMethods;
+  rowSelectionState: UseSelectState;
+};
+
+export const toExternalContext = ({ rowExpand, select }: TableProvideExtra) => ({
+  rowExpand,
+  rowSelect: select[1],
+  rowSelectionState: select[0],
+});
 
 export const TableColumnCollector = createCollector<TableSetupProps, TableColumnSetupProps, TableProvideExtra, true>({
   ...getCollectorOptions('table', true, true),
   tree: true,
 });
-
 
 export type TableColumnCollectorContext = CollectorContext<TableSetupProps, TableColumnSetupProps, TableProvideExtra>;