+++ packages/components/src/components/table/index.ts
@@ -2,4 +2,9 @@ export * from './Table';
 export * from './TableColumn';
 export * from './type';
 export * from './collector';
-export { TABLE_INDEX_COLUMN, tableColumnRendererRegistry } from './TableColumn.renderer';
+export {
+  TABLE_INDEX_COLUMN,
+  tableColumnRendererRegistry,
+  TABLE_RADIO_SELECT_COLUMN,
+  TABLE_CHECKBOX_SELECT_COLUMN,
+} from './TableColumn.renderer';